--- src/algo.ts
@@ -1,13 +1,0 @@
-namespace Uluru {
-
-	export interface algorithm {
-
-		update: (data: ArrayBufferView | string) => algorithm
-
-		finalize: (...args: any) => object
-
-		[key: string]: any
-
-	}
-
-}
