+++ src/algorithm.ts
@@ -1,0 +1,9 @@
+export default interface algorithm {
+
+	update: (data: ArrayBufferView | string) => algorithm
+
+	finalize: (...args: any) => object
+
+	[key: string]: any
+
+}
