+++ src/ascii.ts
@@ -1,35 +1,29 @@
-namespace Uluru {
+import encoding from "./encoding"
 
-	export namespace enc {
+export default class Ascii implements encoding {
 
-		export class Ascii implements encoding {
+	encode(str: string){
 
-			encode(str: string){
+		let u8array = new Uint8Array(str.length)
 
-				let u8array = new Uint8Array(str.length)
+		for(let i = 0, l = str.length; i < l; i++)
+			u8array[i] = str.charCodeAt(i)
 
-				for(let i = 0, l = str.length; i < l; i++)
-					u8array[i] = str.charCodeAt(i)
-	
-				return u8array
+		return u8array
 
-			}
-	
-			decode(bytes: ArrayBufferView){
+	}
 
-				let bytearr = new Uint8Array(bytes.buffer, bytes.byteOffset, bytes.byteLength)
-	
-				let str: string[] = Array(bytearr.length)
+	decode(bytes: ArrayBufferView){
 
-				for(let i = 0, l = bytearr.length; i < l; i++)
-					str[i] = String.fromCharCode(bytearr[i])
+		let bytearr = new Uint8Array(bytes.buffer, bytes.byteOffset, bytes.byteLength)
 
-				return str.join("")
-				
-			}
+		let str: string[] = Array(bytearr.length)
 
-		}
+		for(let i = 0, l = bytearr.length; i < l; i++)
+			str[i] = String.fromCharCode(bytearr[i])
 
+		return str.join("")
+		
 	}
 
 }
