--- src/enc.ts
@@ -1,15 +1,0 @@
-namespace Uluru {
-
-	export namespace enc {
-
-		export interface encoding {
-
-			encode: (str: string) => Uint8Array
-
-			decode: (bytes: ArrayBufferView) => string
-
-		}
-
-	}
-	
-}
