+++ src/encoding.ts
@@ -1,0 +1,7 @@
+export default interface encoding {
+
+	encode: (str: string) => Uint8Array
+
+	decode: (bytes: ArrayBufferView) => string
+
+}
