--- src/extra/export.ts
@@ -1,8 +1,0 @@
-//@ts-ignore
-if(typeof define === "function" && define.amd)
-//@ts-ignore
-	define("Uluru", [],  () => Uluru)
-//@ts-ignore
-if(typeof module != "undefined")
-//@ts-ignore
-	module.exports = Uluru
