--- src/extra/import.ts
@@ -1,11 +1,0 @@
-//@ts-ignore
-var crypto = typeof crypto == "object" ? crypto : typeof global == "object" ? function(){
-	//@ts-ignore
-	var cryp = require("crypto")
-	return {
-		getRandomValues(data: ArrayBufferView){
-			cryp.randomFillSync(data)
-			return data
-		}
-	}
-}() : undefined
