+++ src/index.ts
@@ -1,0 +1,12 @@
+
+export {default as algorithm} from "./algorithm"
+export {default as encoding} from "./encoding"
+
+import Ascii from "./ascii"
+import Base64 from "./base64"
+import Hex from "./hex"
+import Utf8 from "./utf8"
+
+export {Ascii, Base64, Hex, Utf8}
+
+export * from "./easy"
