--- tsconfig.json
@@ -1,31 +1,0 @@
-{
-
-	"compilerOptions" : {
-
-		"outFile" : "uluru.js",
-
-		"target": "ES6",
-
-		"lib": ["DOM", "ESNext"],
-
-		"sourceMap" : false,
-
-		"removeComments": true,
-
-		"declaration": true,
-
-		"watch": false
-
-	},
-
-	"include": [
-
-		"./src/extra/import.ts",
-
-		"./src/*.ts",
-
-		"./src/extra/export.ts"
-
-	],
-
-}
